put everything in cpi.py since pip imports did not work
